fix(providers): add xAI/Grok models to /model picker and docs (#1105)
- Add GROK_FLAGSHIP_ID constant (grok-4.6) and grok alias to DIRECT_MODEL_ALIASES
  so the /model picker accepts and displays grok alongside haiku/sonnet/opus/fable.
  RESERVED_NAMES auto-expands via Object.keys(DIRECT_MODEL_ALIASES).

- Extract shared isValidModelArg helper into src/agent/session/model-validate.ts
  so REPL and Telegram accept exactly the same set of model arguments (xai /
  xai-oauth, openai-compatible, and claude-* wire ids all pass). Previously
  switchModel in info.ts was missing the xai/xai-oauth check present in the
  Telegram surface.

- Update switchModel (REPL /model) to delegate to isValidModelArg instead of
  an inline predicate, eliminating the drift vector.

- Update isValidTelegramModelArg to thin re-export of isValidModelArg.

- Update MODEL_ALIASES_HINT snapshot test to include grok.

- Update website/content/docs/configuration/models-providers.mdx:
  - Add xai/xai-oauth provider section (API-key and OAuth auth modes)
  - Add grok alias row to Available models table
  - Update /model REPL examples to show grok-4.6 and grok
  - Update AFK_MODEL example and AFK_PROVIDER accepted values list
