+++ src/cli/slash/commands/info.ts
@@ -17,15 +17,13 @@ import { formatCost, formatTokens } from '../../format-utils.js';
 import { cacheHitRate } from '../../commands/trace-usage-format.js';
 import { contextLimitFor, MODEL_CONTEXT_LIMITS } from '../../model-limits.js';
 import { renderDebugBanner } from '../../debug-banner.js';
-import { providerForModel } from '../../../agent/providers/index.js';
 import { isModelAvailable } from '../../../agent/auth/model-availability.js';
 import { fetchSubscriptionUsage, type UsageWindow } from '../../../agent/subscription-usage.js';
 import {
   MODEL_ALIASES_HINT,
-  resolveBinding,
-  slotForInput,
   unconfiguredSlotError,
 } from '../../../agent/session/model-slots.js';
+import { isValidModelArg } from '../../../agent/session/model-validate.js';
 import { runPicker } from '../../render/picker.js';
 import type { SlashCommand, SlashContext } from '../types.js';
 import type { AgentModelInput } from '../../../agent/types.js';
@@ -272,19 +270,11 @@ const historyCmd: SlashCommand = {
  * Writes all user-facing output via ctx.out; never throws.
  */
 async function switchModel(ctx: SlashContext, target: string): Promise<void> {
-  // Accept slot tier names / configured custom names, the built-in Claude
-  // identity aliases, a raw Anthropic wire id (e.g. `claude-sonnet-5`), OR full
-  // HF-style org/model ids (routed openai-compatible). Bare unknown strings
-  // (e.g. typos) are rejected — they'd silently fall through to anthropic-direct
-  // and produce an unhelpful API error at turn time.
-  const isKnownAlias = MODEL_ALIASES_HINT.includes(target);
-  const isSlotName = slotForInput(target) !== undefined;
-  const isOpenAICompatibleId = providerForModel(target) === 'openai-compatible';
-  // The resolved id's `claude-` prefix is the confident raw-Claude-id signal;
-  // aliases already resolve to a claude- id too, so this also covers them.
-  const resolvedId = resolveBinding(target).id.trim().toLowerCase();
-  const isClaudeWireId = resolvedId.startsWith('claude-') || resolvedId.startsWith('claude_');
-  if (!isKnownAlias && !isSlotName && !isOpenAICompatibleId && !isClaudeWireId) {
+  // Accept slot tier names / configured custom names, the built-in identity
+  // aliases (claude-*, grok-*, gpt-*, org/model HF ids, …). Bare unknown
+  // strings (e.g. typos) are rejected — they'd silently fall through to a
+  // provider and produce an unhelpful API error at turn time.
+  if (!isValidModelArg(target)) {
     ctx.out.warn(`Unknown model: ${target}. Aliases: ${MODEL_ALIASES_HINT.join(', ')}  (or a full model id)`);
     return;
   }
