+++ src/telegram/handlers/commands.model-validate.ts
@@ -1,25 +1,16 @@
 /**
  * Accept/reject matrix for Telegram `/model <id>` (aliases, slots, wire ids).
  *
+ * Delegates to the shared {@link isValidModelArg} helper so REPL and Telegram
+ * cannot drift — both surfaces accept exactly the same set of model arguments.
+ *
  * @module telegram/handlers/commands.model-validate
  */
 
-import { providerForModel } from '../../agent/providers/index.js';
-import {
-  MODEL_ALIASES_HINT,
-  resolveBinding,
-  slotForInput,
-} from '../../agent/session/model-slots.js';
+import { isValidModelArg } from '../../agent/session/model-validate.js';
 import type { AgentModelInput } from '../../agent/types.js';
 
 /** True when `model` is a known alias, slot name, or routable full wire id. */
 export function isValidTelegramModelArg(model: AgentModelInput): boolean {
-  if (MODEL_ALIASES_HINT.includes(model)) return true;
-  if (slotForInput(model) !== undefined) return true;
-  const routed = providerForModel(model);
-  if (routed === 'openai-compatible') return true;
-  if (routed === 'xai' || routed === 'xai-oauth') return true;
-  // Raw Anthropic wire id (e.g. `claude-sonnet-5`).
-  const resolvedId = resolveBinding(model).id.trim().toLowerCase();
-  return resolvedId.startsWith('claude-') || resolvedId.startsWith('claude_');
+  return isValidModelArg(model);
 }
